+++ Toolkit/Tests/e2e/Get-RscSlaDomain.Tests.ps1
@@ -3,15 +3,15 @@ BeforeAll {
 }
 
 # Broken: replace -Skip for -Fixture to run the tests
-Describe -Name 'Get-RscSlaDomain' -Tag 'Public' -Skip {
+Describe -Name 'Get-RscSla' -Tag 'Public' -Skip {
     Context -Name 'Id' {
         It -Name 'retrieves single SLA Domain' -Test {
-            { Get-RscSlaDomain -Id '00000000-0000-0000-0000-000000000000' | Out-Null} | Should -Not -Throw
+            { Get-RscSla -Id '00000000-0000-0000-0000-000000000000' | Out-Null} | Should -Not -Throw
         }
     }
     Context -Name 'Name' {
         It -Name 'retrieves multiple SLA Domains' -Test {
-            { Get-RscSlaDomain | Out-Null} | Should -Not -Throw
+            { Get-RscSla | Out-Null} | Should -Not -Throw
         }
     }
 }